Switched db to .ts file
